Finished M5git add .! Added permissions to delete

+++ backend/src/resolvers/Mutation.js
@@ -50,9 +50,16 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     //1. Find the item
-    const item = await ctx.db.query.item({ where }, `{id title}`);
+    const item = await ctx.db.query.item({ where }, `{id title user{id}}`);
     //2. Check if they own the item, or have permissions
-    //TODO
+    const ownsItem = item.user.id === ctx.request.userId;
+    console.log(ctx.request.user.permissions);
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission),
+    );
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permission to do that");
+    }
     //3. Delete
     return ctx.db.mutation.deleteItem({ where }, info);
   },
